Slight restructure to the binary installer and add the rsync updater

+++ makebinpkg.sh
@@ -1,22 +1,20 @@
-cd oolite.app/Contents/Resources
-rm -rf .svn
-rm -rf AIs/.svn
-rm -rf Config/.svn
-rm -rf Images/.svn
-rm -rf Models/.svn
-rm -rf Music/.svn
-rm -rf Sounds/.svn
-rm -rf Textures/.svn
-cd ../../..
+#!/bin/sh
+if [ ! $1 ]
+then
+   echo "Usage: makebinpkg.sh <releasename>"
+   echo
+   exit
+fi
 
 rm -rf $HOME/oolite-installer
 mkdir -p $HOME/oolite-installer
-tar cvf ~/oolite-installer/oolite-app.tar oolite.app
+tar cvf ~/oolite-installer/oolite-app.tar oolite.app --exclude .svn
 
 cd SelfContainedInstaller
-cp install README.TXT PLAYING.TXT oolite ~/oolite-installer
-tar cvf ~/oolite-installer/oolite-deps.tar oolite-deps
+cp install oolite-update README.TXT PLAYING.TXT FAQ.TXT oolite ~/oolite-installer
+tar cvf ~/oolite-installer/oolite-deps.tar oolite-deps --exclude .svn
+echo $1 >~/oolite-installer/release.txt
 
 cd ~/
-tar zcvf oolite-x86-installer-`date +%Y%m%d`.tar.gz oolite-installer
+tar zcvf oolite-x86-installer-$1.tar.gz oolite-installer
